modified the query in repository. Changed Long to int in parameter list for getMeal

+++ src/main/java/com/techreturners/teama/healthyfood/api/controller/HealthyFoodController.java
@@ -45,7 +45,7 @@ public class HealthyFoodController {
 
     @GetMapping({"/meal"})
     @Operation(summary = "Gets a selection of meals based on restrictions")
-    public ResponseEntity<List<Meal>> getMeals(@RequestParam Long calories, @RequestParam List<String> excludedIngredients, @RequestParam List<String> diet, String category) {
+    public ResponseEntity<List<Meal>> getMeals(@RequestParam int calories, @RequestParam List<String> excludedIngredients, @RequestParam List<String> diet, String category) {
         List<Meal> meals = healthyFoodManagerService.getMeals(calories, excludedIngredients, diet, category);
         return new ResponseEntity<>(meals, HttpStatus.OK);
     }
